Part 2 of massive AccuracyEstimator change
I believe that all of the internal changes are done. The settings popout works correctly and interacts with the embedded AccuracyEstimator objects as expected, and all of the data needed for the GIFs is collected and ready for usage. All I need to do is write the actual Visualizer :tada: