Updated Comparator to let users load builds into columns from the base GUI
I noticed a major UX flaw in my original idea -- even I, the creator of this program, disliked having to manually enter in all the build Strings using correct syntax and capitalization. Additionally, the popup covers the vast majority of the buttons so it's hard to figure it out on the fly. Hopefully this change will fix that issue.