First pass at smoothing out MultiLineGraph
Did a super-basic rolling average that uses up to 2 neighbors on either side of a data point to get the big jumps taken care of, and then played a little with the BasicStroke settings to look for some under-the-hood tweaks that were available. It definitely looks a little better now.