Fixed a bug where averageTimeToIgnite() could add negative or infinite ignite times per enemy
PGL builds are able to deal less heat/sec than Oppressor's CoolingRate. This created situations where it took an Infinity amount of time to ignite (divide by zero), or a negative time to ignite.

Technically this is still reducing the average time to ignite, but it's on purpose instead of because the code is adding a negative ignition time.