Removed source code for defunct Export Mods/OCs functionality
Used it once when Karl.gg was made, and never since. This is primarily due to the Karl.gg devs deciding to not track mod/OC revision history. No user complained of the menuitems' removal in 1.1.0, so I'm pulling the plug and removing the code too. Hopefully this decreases some bloat in the codebase.