Refactored the three Weapon "Explanation" methods to be dynamic
These methods threw a bunch of errors when I added TriJaw and Brundle in U33. As an extension of the new Enemy object, I'm able to imitate the iterative logic of breakpoints, armorwasting, and overkill. This means that these new explanation methods will automatically update with new enemy objects that get added in the future, or flags or properties that get changed in the future.