Fixed a bug in Breakpoints where Brundle's Armor could be bypassed
If the armor damage per shot was greater than the eHP of the Brundle's Armor plate, the numShotsToBreak would be less than 1. Thus, the first shot would skip the armor entirely, even with an AB <= 100%. By changing it to use `ceil()`, now an AB mod is required to achieve that effect (as intended).