Reduced RectangularAccuracy Precision from 100 to 50
This made Autocannon 3 sec faster per metric, Boomstick 2.7 sec faster, and Zhukovs 1.8 sec faster. Now AC is still the longest to calculate, but only by a factor of 2.5 and it's primarily due to the number of mods and the Area Damage shenanigans.

Suggested by Fabricio in DCT.

Also: fixed some numbers in Sludge Puddle DoT and Web Spitter base HP.