Moved where General and Weakpoint Accuracies get multiplied by 100
Before it happened at base level, and then I had to divide by 100 all of the place in the model. It really got on my nerves, so now the model is beautiful code and the only place where the x100 happens is user-facing. Things like GUI, MetricCalculator, and the Comparators.